add test assets
