+++ endoreg_db/models/readme.md
@@ -1,5 +1,10 @@
 # DevLog
+## Testing
+Run individual test submodules like this:
+`python runtests.py "media/video"`
 
+Split Videos using video Splitter script:
+`python endoreg-db/scripts/video_splitter.py --input ./endoreg-db/data/raw_videos/lux-gastro-video.mp4 --interval 15`
 # EndoReg DB Models
 Summary by submodules.
