+++ endoreg_db/utils/video/__init__.py
@@ -2,6 +2,7 @@ from .transcode_videofile import (
     transcode_videofile,
     transcode_videofile_if_required,
 )
+from .video_splitter import split_video
 
 from .extract_frames import extract_frames, initialize_frame_objects
 
@@ -10,4 +11,5 @@ __all__ = [
     "initialize_frame_objects",
     "transcode_videofile",
     "transcode_videofile_if_required",
+    "split_video",
 ]
